+++ src/main/java/algorithms/IntegrationMethod.java
@@ -2,5 +2,5 @@ package algorithms;
 
 public interface IntegrationMethod {
 
-	double updatePosition(double dt);
+	float updatePosition(float dt);
 }
