+++ src/main/java/algorithms/TimeStepDrivenMolecularDynamics.java
@@ -20,29 +20,29 @@ public class TimeStepDrivenMolecularDynamics {
 	};
 
 	// Initial State
-	private static double time = 0.0;
+	private static float time = 0.0F;
 
 	public static void run(
-			double limitTime,
-			double dt,
-			double printDeltaT,
-			double k,
-			double gamma,
-			double initialPosition,
-			double mass,
+			float limitTime,
+			float dt,
+			float printDeltaT,
+			float k,
+			float gamma,
+			float initialPosition,
+			float mass,
 			String positionsPlotFile) {
 
 		// Initialize state
-		double initialVelocity = -1 * (gamma / (2 * mass)); // m/s (initial velocity at time=0)
+		float initialVelocity = -1 * (gamma / (2 * mass)); // m/s (initial velocity at time=0)
 
-		Stack<Double> timeStepValues = new Stack<>();
-		Stack<Double> analyticValues = new Stack<>();
-		Stack<Double> beemanPositionValues = new Stack<>();
-		Stack<Double> velocityVerletPositionValues = new Stack<>();
-		Stack<Double> order5GearPredictorCorrectorPositionValues = new Stack<>();
+		Stack<Float> timeStepValues = new Stack<>();
+		Stack<Float> analyticValues = new Stack<>();
+		Stack<Float> beemanPositionValues = new Stack<>();
+		Stack<Float> velocityVerletPositionValues = new Stack<>();
+		Stack<Float> order5GearPredictorCorrectorPositionValues = new Stack<>();
 
 		Force springForce = new SpringForce(k, gamma);
-		double previousAcceleration = springForce.F(initialPosition, initialVelocity) / mass;
+		float previousAcceleration = springForce.F(initialPosition, initialVelocity) / mass;
 
 		SpringAnalyticSolution springAnalyticSolution = new SpringAnalyticSolution();
 		Beeman beeman = new Beeman(mass, initialPosition, initialVelocity, previousAcceleration, springForce);
@@ -56,7 +56,7 @@ public class TimeStepDrivenMolecularDynamics {
 		order5GearPredictorCorrectorPositionValues.push(initialPosition);
 
 		// MSE for each integrator
-		double[] mse = new double[3];
+		float[] mse = new float[3];
 
 		// Print frame
 		int currentFrame = 1;
@@ -65,15 +65,15 @@ public class TimeStepDrivenMolecularDynamics {
 		while (time < limitTime) {
 			time += dt;
 
-			double analyticValue = springAnalyticSolution.getPosition(k, gamma, mass, time);
+			float analyticValue = springAnalyticSolution.getPosition(k, gamma, mass, time);
 
-			double beemanPosition = beeman.updatePosition(dt);
+			float beemanPosition = beeman.updatePosition(dt);
 			mse[0] += Math.pow(analyticValue - beemanPosition, 2);
 
-			double velocityPosition = velocityVerlet.updatePosition(dt);
+			float velocityPosition = velocityVerlet.updatePosition(dt);
 			mse[1] += Math.pow(analyticValue - velocityPosition, 2);
 
-			double order5GearPredictorCorrectorPosition = order5GearPredictorCorrector.updatePosition(dt);
+			float order5GearPredictorCorrectorPosition = order5GearPredictorCorrector.updatePosition(dt);
 			mse[2] += Math.pow(analyticValue - order5GearPredictorCorrectorPosition, 2);
 
 			if ((currentFrame % printFrame) == 0) {
